Added static libraries and began adding bilinear interpolation code.
Static libraries are now included because they are small, and are not available in the repositories for the target Raspbian system.
The bilinear code is incomplete. It is unknown whether I will use Yeppp! armhf to do the trig calculations in vector format. For simplicity, I dont think I will at the moment.