Completed debugging of the program.
Radius did not account for the width of the image, making it far too small.
The coordinates were not centered on the image and were negative.
Added assert to catch negative coordinates to interpolate.
Image dimension were not squared to calculate new size (was XOR instead).
Slightly optimized generateSlice (although compiler would probably cover it).